Added unit test for nonprod population load.
